Release 4 (APCS-140)

See merge request apcs/kb-server!51

+++ src/main/java/com/nlmk/kb/server/service/result/configuration/ResultConfigService.java
@@ -12,14 +12,17 @@ public interface ResultConfigService {
 
     Page<ResultsConfigDto> findPyPage(PageRequest pageRequest);
 
-    ResultsConfigDto findById(Long id);
+    ResultsConfigDto findById(Integer id);
 
     ResultsConfigDto create(@NotNull @Valid ResultsConfigDto dto);
 
     ResultsConfigDto update(@NotNull @Valid ResultsConfigDto dto);
 
+    /**
+     * Получение всех активных конфигураций
+     */
     List<ResultsConfigDto> getEnabledTopics();
 
-    void deleteById(long id);
+    void deleteById(Integer id);
 
 }
